% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/circles.R
\name{getCodes}
\alias{getCodes}
\title{Extract available CWatM circles}
\usage{
getCodes(flux = NULL)
}
\arguments{
\item{flux}{A `character` vector with flux codes. See `getFluxes()`.}
}
\value{
a `data.frame` with available circles per flux.
}
\description{
The function extracts the available circles' codes for  any selected flux.
}
\examples{
getCodes(flux = "hydro")

}
