% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/circles.R
\name{getFluxes}
\alias{getFluxes}
\title{Extract available CWatM fluxes}
\usage{
getFluxes()
}
\value{
a `character` vector of  fluxes.
}
\description{
The function extracts the available fluxes codes from the CWatM Circles database.
}
\examples{
getFluxes()

}
