% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/circles.R
\name{getVars}
\alias{getVars}
\title{Extract required CWatM circle variables}
\usage{
getVars(
  circle = NULL,
  flux = "hydro",
  modflow = FALSE,
  wastewater = FALSE,
  limitAbstr = FALSE
)
}
\arguments{
\item{circle}{A `character` vector with circle names. See `getCodes()`, (default: `NULL`, so all circles are included).}

\item{flux}{A `character` vector with flux codes. See `getFluxes()`, (default: `hydro`, so only water-circles are included).}

\item{modflow}{A `logical` variable indicating if modflow simulation was included (default: `FALSE`).}

\item{wastewater}{A `logical` variable indicating if the wastewater module was included (default: `FALSE`).}

\item{limitAbstr}{A `logical` variable indicating if limitAbstraction was set to `TRUE`, (default: `FALSE`).}
}
\value{
a comma-separated `character` string with all required output variables.
}
\description{
The function provide a string with all  CWatM output variables, required to create  selected circles, given specified settings.
}
\examples{
getVars(circle = c("soil", "main"), flux = "hydro", limitAbstr = TRUE)

}
